requirements.txt
django>=5.2.3
psycopg2-binary
python-decouple
djangorestframework
drf-yasg
django-filter
faker
gunicorn  